[ANDROSDK-904] Class visibility
