+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/internal/TrackedEntityInstanceService.java
@@ -34,7 +34,7 @@ import org.hisp.dhis.android.core.arch.api.payload.internal.Payload;
 import org.hisp.dhis.android.core.imports.internal.HttpMessageResponse;
 import org.hisp.dhis.android.core.imports.internal.TEIWebResponse;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstance;
-import org.hisp.dhis.android.core.trackedentity.search.internal.SearchGrid;
+import org.hisp.dhis.android.core.trackedentity.search.SearchGrid;
 
 import java.util.List;
