+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/search/SearchGridMapper.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.trackedentity.search.internal;
+package org.hisp.dhis.android.core.trackedentity.search;
 
 import org.hisp.dhis.android.core.common.BaseIdentifiableObject;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttributeValue;
@@ -49,7 +49,7 @@ class SearchGridMapper {
         // Empty constructor to add Dagger annotation
     }
 
-    public List<TrackedEntityInstance> transform(SearchGrid searchGrid) throws ParseException {
+    List<TrackedEntityInstance> transform(SearchGrid searchGrid) throws ParseException {
         List<TrackedEntityInstance> teis = new ArrayList<>(searchGrid.rows().size());
         for (List<String> row : searchGrid.rows()) {
             TrackedEntityInstance tei = TrackedEntityInstance.builder()
