+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/search/TrackedEntityInstanceQueryRepositoryScope.java
@@ -26,28 +26,31 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.trackedentity.search.internal;
+package org.hisp.dhis.android.core.trackedentity.search;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.google.auto.value.AutoValue;
 
-import org.hisp.dhis.android.core.arch.call.queries.internal.BaseQuery;
 import org.hisp.dhis.android.core.arch.dateformat.internal.SafeDateFormat;
+import org.hisp.dhis.android.core.arch.repositories.scope.BaseScope;
+import org.hisp.dhis.android.core.arch.repositories.scope.internal.RepositoryMode;
 import org.hisp.dhis.android.core.arch.repositories.scope.internal.RepositoryScopeFilterItem;
+import org.hisp.dhis.android.core.common.State;
 import org.hisp.dhis.android.core.organisationunit.OrganisationUnitMode;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @AutoValue
-public abstract class TrackedEntityInstanceQueryOnline extends BaseQuery {
+abstract class TrackedEntityInstanceQueryRepositoryScope implements BaseScope {
 
     private static final SafeDateFormat QUERY_FORMAT = new SafeDateFormat("yyyy-MM-dd");
+
+    @NonNull
+    public abstract RepositoryMode mode();
 
     @NonNull
     public abstract List<String> orgUnits();
@@ -59,13 +62,13 @@ public abstract class TrackedEntityInstanceQueryOnline extends BaseQuery {
     public abstract String program();
 
     @Nullable
-    public abstract String query();
+    public abstract RepositoryScopeFilterItem query();
 
-    @Nullable
-    public abstract List<String> attribute();
+    @NonNull
+    public abstract List<RepositoryScopeFilterItem> attribute();
 
-    @Nullable
-    public abstract List<String> filter();
+    @NonNull
+    public abstract List<RepositoryScopeFilterItem> filter();
 
     @Nullable
     public abstract Date programStartDate();
@@ -76,76 +79,51 @@ public abstract class TrackedEntityInstanceQueryOnline extends BaseQuery {
     @Nullable
     public abstract String trackedEntityType();
 
-    @NonNull
+    @Nullable
     public abstract Boolean includeDeleted();
 
-    String formattedProgramStartDate() {
+    @Nullable
+    public abstract List<State> states();
+
+    public String formattedProgramStartDate() {
         return programStartDate() == null ? null : QUERY_FORMAT.format(programStartDate());
     }
 
-    String formattedProgramEndDate() {
+    public String formattedProgramEndDate() {
         return programEndDate() == null ? null : QUERY_FORMAT.format(programEndDate());
     }
 
     public abstract Builder toBuilder();
 
     public static Builder builder() {
-        return new AutoValue_TrackedEntityInstanceQueryOnline.Builder();
+        return new AutoValue_TrackedEntityInstanceQueryRepositoryScope.Builder()
+                .attribute(Collections.emptyList())
+                .filter(Collections.emptyList())
+                .orgUnits(Collections.emptyList())
+                .mode(RepositoryMode.OFFLINE_ONLY)
+                .includeDeleted(false);
     }
 
-    public static TrackedEntityInstanceQueryOnline create(TrackedEntityInstanceQueryRepositoryScope scope) {
-        String query = null;
-        if (scope.query() != null) {
-            query = scope.query().operator().getApiUpperOperator() + ":" + scope.query().value();
-        }
-
-        return TrackedEntityInstanceQueryOnline.builder()
-                .query(query)
-                .attribute(toAPIFilterFormat(scope.attribute()))
-                .filter(toAPIFilterFormat(scope.filter()))
-                .orgUnits(scope.orgUnits())
-                .orgUnitMode(scope.orgUnitMode())
-                .program(scope.program())
-                .programStartDate(scope.programStartDate())
-                .programEndDate(scope.programEndDate())
-                .trackedEntityType(scope.trackedEntityType())
-                .includeDeleted(false)
-                .page(1)
-                .pageSize(50)
-                .paging(true)
-                .build();
-    }
-
-    private static List<String> toAPIFilterFormat(List<RepositoryScopeFilterItem> items) {
-        Map<String, String> itemMap = new HashMap<>();
-        for (RepositoryScopeFilterItem item : items) {
-            String filterClause = ":" + item.operator().getApiUpperOperator() + ":" + item.value();
-            String existingClause = itemMap.get(item.key());
-            String newClause = (existingClause == null ? "" : existingClause) + filterClause;
-
-            itemMap.put(item.key(), newClause);
-        }
-
-        List<String> itemList = new ArrayList<>();
-        for (Map.Entry<String, String> entry : itemMap.entrySet()) {
-            itemList.add(entry.getKey() + entry.getValue());
-        }
-        return itemList;
+    public static TrackedEntityInstanceQueryRepositoryScope empty() {
+        return builder().build();
     }
 
     @AutoValue.Builder
-    public abstract static class Builder extends BaseQuery.Builder<Builder> {
+    abstract static class Builder {
+
+        public abstract Builder mode(RepositoryMode mode);
+
         public abstract Builder orgUnits(List<String> orgUnits);
 
         public abstract Builder orgUnitMode(OrganisationUnitMode orgUnitMode);
 
         public abstract Builder program(String program);
 
-        public abstract Builder query(String query);
+        public abstract Builder query(RepositoryScopeFilterItem query);
 
-        public abstract Builder attribute(List<String> attribute);
+        public abstract Builder attribute(List<RepositoryScopeFilterItem> attribute);
 
-        public abstract Builder filter(List<String> filter);
+        public abstract Builder filter(List<RepositoryScopeFilterItem> filter);
 
         public abstract Builder programStartDate(Date programStartDate);
 
@@ -153,8 +131,21 @@ public abstract class TrackedEntityInstanceQueryOnline extends BaseQuery {
 
         public abstract Builder trackedEntityType(String trackedEntityType);
 
+        public abstract Builder states(List<State> states);
+
         public abstract Builder includeDeleted(Boolean includeDeleted);
 
-        public abstract TrackedEntityInstanceQueryOnline build();
+        abstract TrackedEntityInstanceQueryRepositoryScope autoBuild();
+
+        // Auxiliary fields to access values
+        abstract List<State> states();
+
+        public TrackedEntityInstanceQueryRepositoryScope build() {
+            if (states() != null) {
+                mode(RepositoryMode.OFFLINE_ONLY);
+            }
+
+            return autoBuild();
+        }
     }
 }
